requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Werkzeug==3.1.3
colorama==0.4.6
flask_sqlalchemy==3.1.1
greenlet==3.2.1
sqlalchemy==2.0.40
typing_extensions==4.13.2